tidy codebase and split up routes
